move wow to different .py file
